Fixing replacement of sly based parser
