+++ pygeofilter/parsers/ecql/__init__.py
@@ -1,4 +1,1 @@
-# https://docs.opengeospatial.org/is/12-168r6/12-168r6.html#54
-# https://github.com/geotools/geotools/blob/main/modules/library/cql/ECQL.md
-
 from .parser import parse
