--- pygeofilter/parsers/ecql/ecql_lark/__init__.py
@@ -1,1 +1,0 @@
-from .parser import parse
